Updating namespace
