+++ NuGetTransitiveDependencyFinder.sln
@@ -6,9 +6,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28701.123
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGet.TransitiveDependency.Finder", "src\NuGet.TransitiveDependency.Finder\NuGet.TransitiveDependency.Finder.csproj", "{FAEC5437-7DF8-4DFB-872A-6BEBF216A3F4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGetTransitiveDependencyFinder", "src\NuGetTransitiveDependencyFinder\NuGetTransitiveDependencyFinder.csproj", "{FAEC5437-7DF8-4DFB-872A-6BEBF216A3F4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGet.TransitiveDependency.Finder.ConsoleApp", "src\NuGet.TransitiveDependency.Finder.ConsoleApp\NuGet.TransitiveDependency.Finder.ConsoleApp.csproj", "{B128460D-E702-45EA-B6B4-FAF8C53E1456}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGetTransitiveDependencyFinder.ConsoleApp", "src\NuGetTransitiveDependencyFinder.ConsoleApp\NuGetTransitiveDependencyFinder.ConsoleApp.csproj", "{B128460D-E702-45EA-B6B4-FAF8C53E1456}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
