+++ src/NuGetTransitiveDependencyFinder.ConsoleApp/Program.cs
@@ -3,12 +3,12 @@
 // Licensed under the MIT License
 // </copyright>
 
-namespace NuGet.TransitiveDependency.Finder.ConsoleApp
+namespace NuGetTransitiveDependencyFinder.ConsoleApp
 {
     using Microsoft.Extensions.Logging;
     using Microsoft.Extensions.Logging.Console;
-    using NuGet.TransitiveDependency.Finder.ConsoleApp.Output;
-    using NuGet.TransitiveDependency.Finder.ConsoleApp.Resources;
+    using NuGetTransitiveDependencyFinder.ConsoleApp.Output;
+    using NuGetTransitiveDependencyFinder.ConsoleApp.Resources;
 
     /// <summary>
     /// The main class of the application, defining the entry point and all interaction.
