+++ src/NuGetTransitiveDependencyFinder/NuGetTransitiveDependencyFinder.csproj
