+++ src/NuGetTransitiveDependencyFinder/Output/Framework.cs
@@ -1,25 +1,29 @@
-// <copyright file="Project.cs" company="Muiris Woulfe">
+// <copyright file="Framework.cs" company="Muiris Woulfe">
 // © Muiris Woulfe
 // Licensed under the MIT License
 // </copyright>
 
-namespace NuGet.TransitiveDependency.Finder.Output
+namespace NuGetTransitiveDependencyFinder.Output
 {
     using System;
+    using NuGet.Frameworks;
 
     /// <summary>
-    /// A class representing the outputted project information.
+    /// A class representing the outputted .NET framework information for each project.
     /// </summary>
-    /// <remarks>The child elements of this class are the <see cref="Framework"/> objects.</remarks>
-    public sealed class Project : IdentifiedBase<string, Framework>, IComparable<Project>, IEquatable<Project>
+    /// <remarks>The child elements of this class are the transitive NuGet dependencies.</remarks>
+    public sealed class Framework :
+        IdentifiedBase<NuGetFramework, string>,
+        IComparable<Framework>,
+        IEquatable<Framework>
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="Project"/> class.
+        /// Initializes a new instance of the <see cref="Framework"/> class.
         /// </summary>
-        /// <param name="capacity">The quantity of .NET framework identifiers for which the collection initially has
+        /// <param name="capacity">The quantity of transitive NuGet dependencies for which the collection initially has
         /// adequate capacity.</param>
-        /// <param name="identifier">The project identifier.</param>
-        public Project(int capacity, string identifier)
+        /// <param name="identifier">The .NET framework identifier.</param>
+        public Framework(int capacity, NuGetFramework identifier)
             : base(capacity, identifier)
         {
         }
@@ -33,7 +37,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is equal to <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator ==(Project left, Project right) =>
+        public static bool operator ==(Framework left, Framework right) =>
             left.CompareTo(right) == 0;
 
         /// <summary>
@@ -45,7 +49,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is not equal to <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator !=(Project left, Project right) =>
+        public static bool operator !=(Framework left, Framework right) =>
             left.CompareTo(right) != 0;
 
         /// <summary>
@@ -57,7 +61,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is less than <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator <(Project left, Project right) =>
+        public static bool operator <(Framework left, Framework right) =>
             left.CompareTo(right) < 0;
 
         /// <summary>
@@ -69,7 +73,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is less than or equal to <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator <=(Project left, Project right) =>
+        public static bool operator <=(Framework left, Framework right) =>
             left.CompareTo(right) <= 0;
 
         /// <summary>
@@ -81,7 +85,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is greater than <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator >(Project left, Project right) =>
+        public static bool operator >(Framework left, Framework right) =>
             left.CompareTo(right) > 0;
 
         /// <summary>
@@ -93,19 +97,19 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is greater than or equal to <see paramref="right"/>;
         /// otherwise, <c>false</c>.</returns>
-        public static bool operator >=(Project left, Project right) =>
+        public static bool operator >=(Framework left, Framework right) =>
             left.CompareTo(right) >= 0;
 
         /// <inheritdoc/>
         /// <remarks>The result of this method is solely dependent on
         /// <see cref="IdentifiedBase{TIdentifier, TChild}.Identifier"/>.</remarks>
-        public int CompareTo(Project? other) =>
+        public int CompareTo(Framework? other) =>
             this.BaseCompareTo(other);
 
         /// <inheritdoc/>
         /// <remarks>The result of this method is solely dependent on
         /// <see cref="IdentifiedBase{TIdentifier, TChild}.Identifier"/>.</remarks>
-        public bool Equals(Project? other) =>
+        public bool Equals(Framework? other) =>
             this.BaseCompareTo(other) == 0;
 
         /// <inheritdoc/>
@@ -121,7 +125,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
             this.BaseHashCode;
 
         /// <inheritdoc/>
-        protected override bool IsAddValid(Framework child) =>
-            child.HasChildren;
+        protected override bool IsAddValid(string child) =>
+            true;
     }
 }
