+++ src/NuGetTransitiveDependencyFinder/Output/Project.cs
@@ -1,29 +1,25 @@
-// <copyright file="Framework.cs" company="Muiris Woulfe">
+// <copyright file="Project.cs" company="Muiris Woulfe">
 // © Muiris Woulfe
 // Licensed under the MIT License
 // </copyright>
 
-namespace NuGet.TransitiveDependency.Finder.Output
+namespace NuGetTransitiveDependencyFinder.Output
 {
     using System;
-    using NuGet.Frameworks;
 
     /// <summary>
-    /// A class representing the outputted .NET framework information for each project.
+    /// A class representing the outputted project information.
     /// </summary>
-    /// <remarks>The child elements of this class are the transitive NuGet dependencies.</remarks>
-    public sealed class Framework :
-        IdentifiedBase<NuGetFramework, string>,
-        IComparable<Framework>,
-        IEquatable<Framework>
+    /// <remarks>The child elements of this class are the <see cref="Framework"/> objects.</remarks>
+    public sealed class Project : IdentifiedBase<string, Framework>, IComparable<Project>, IEquatable<Project>
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="Framework"/> class.
+        /// Initializes a new instance of the <see cref="Project"/> class.
         /// </summary>
-        /// <param name="capacity">The quantity of transitive NuGet dependencies for which the collection initially has
+        /// <param name="capacity">The quantity of .NET framework identifiers for which the collection initially has
         /// adequate capacity.</param>
-        /// <param name="identifier">The .NET framework identifier.</param>
-        public Framework(int capacity, NuGetFramework identifier)
+        /// <param name="identifier">The project identifier.</param>
+        public Project(int capacity, string identifier)
             : base(capacity, identifier)
         {
         }
@@ -37,7 +33,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is equal to <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator ==(Framework left, Framework right) =>
+        public static bool operator ==(Project left, Project right) =>
             left.CompareTo(right) == 0;
 
         /// <summary>
@@ -49,7 +45,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is not equal to <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator !=(Framework left, Framework right) =>
+        public static bool operator !=(Project left, Project right) =>
             left.CompareTo(right) != 0;
 
         /// <summary>
@@ -61,7 +57,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is less than <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator <(Framework left, Framework right) =>
+        public static bool operator <(Project left, Project right) =>
             left.CompareTo(right) < 0;
 
         /// <summary>
@@ -73,7 +69,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is less than or equal to <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator <=(Framework left, Framework right) =>
+        public static bool operator <=(Project left, Project right) =>
             left.CompareTo(right) <= 0;
 
         /// <summary>
@@ -85,7 +81,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is greater than <see paramref="right"/>; otherwise,
         /// <c>false</c>.</returns>
-        public static bool operator >(Framework left, Framework right) =>
+        public static bool operator >(Project left, Project right) =>
             left.CompareTo(right) > 0;
 
         /// <summary>
@@ -97,19 +93,19 @@ namespace NuGet.TransitiveDependency.Finder.Output
         /// <param name="right">The right operand to compare.</param>
         /// <returns><c>true</c> if <see paramref="left"/> is greater than or equal to <see paramref="right"/>;
         /// otherwise, <c>false</c>.</returns>
-        public static bool operator >=(Framework left, Framework right) =>
+        public static bool operator >=(Project left, Project right) =>
             left.CompareTo(right) >= 0;
 
         /// <inheritdoc/>
         /// <remarks>The result of this method is solely dependent on
         /// <see cref="IdentifiedBase{TIdentifier, TChild}.Identifier"/>.</remarks>
-        public int CompareTo(Framework? other) =>
+        public int CompareTo(Project? other) =>
             this.BaseCompareTo(other);
 
         /// <inheritdoc/>
         /// <remarks>The result of this method is solely dependent on
         /// <see cref="IdentifiedBase{TIdentifier, TChild}.Identifier"/>.</remarks>
-        public bool Equals(Framework? other) =>
+        public bool Equals(Project? other) =>
             this.BaseCompareTo(other) == 0;
 
         /// <inheritdoc/>
@@ -125,7 +121,7 @@ namespace NuGet.TransitiveDependency.Finder.Output
             this.BaseHashCode;
 
         /// <inheritdoc/>
-        protected override bool IsAddValid(string child) =>
-            true;
+        protected override bool IsAddValid(Framework child) =>
+            child.HasChildren;
     }
 }
