+++ src/NuGetTransitiveDependencyFinder/Output/Projects.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License
 // </copyright>
 
-namespace NuGet.TransitiveDependency.Finder.Output
+namespace NuGetTransitiveDependencyFinder.Output
 {
     /// <summary>
     /// A class representing the outputted projects information.
