+++ src/NuGetTransitiveDependencyFinder/ProjectAnalysis/Assets.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License
 // </copyright>
 
-namespace NuGet.TransitiveDependency.Finder.ProjectAnalysis
+namespace NuGetTransitiveDependencyFinder.ProjectAnalysis
 {
     using System.IO;
     using Microsoft.Extensions.Logging;
